move partials logics to its own file
